backend/requirements.txt
flask
flask-cors
Werkzeug
python-dotenv
langchain
google-cloud-aiplatform
google-cloud-storage
beautifulsoup4
requests
pinecone
pinecone-client
PyPDF2
PyMuPDF
<<<<<<< Updated upstream
pytest
=======
pytest
langchain-pinecone 
pinecone-notebooks
spire
spire.pdf
>>>>>>> Stashed changes
